Move TestDisks to separate file, refactor test methods
