+++ circletracking/tests/test_locate.py
@@ -30,7 +30,7 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
         radius_fit, center_fit, _ = fit_ellipse(points, mode='xy')
 
         return np.concatenate([radius_fit, center_fit]), \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
 
     @repeat_test_std
@@ -46,7 +46,7 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
         radius_fit, center_fit, _ = fit_ellipse(points, mode='xy')
 
         return np.concatenate([radius_fit, center_fit]), \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
 
     @repeat_test_std
@@ -62,7 +62,7 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
         radius_fit, center_fit, _ = fit_ellipse(points, mode='0')
 
         return np.concatenate([radius_fit, center_fit]), \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
 
     @repeat_test_std
@@ -78,7 +78,7 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
         radius_fit, center_fit, _ = fit_ellipse(points, mode='0')
 
         return np.concatenate([radius_fit, center_fit]), \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
     #
     # def test_fit_ellipse(self):
@@ -99,15 +99,15 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
     #     points += center[:, np.newaxis]
     #     #points += (np.random.random(points.shape) - 0.5) * noise
 
-        # radius_fit, center_fit, angle_fit = gt.fit_ellipse(points, mode='')
-        # assert_allclose(radius, radius_fit, atol=0.1)
-        # assert_allclose(center, center_fit, atol=0.1)
-        # assert_allclose(angle, angle_fit, atol=0.1)
+    # radius_fit, center_fit, angle_fit = gt.fit_ellipse(points, mode='')
+    # assert_allclose(radius, radius_fit, atol=0.1)
+    # assert_allclose(center, center_fit, atol=0.1)
+    # assert_allclose(angle, angle_fit, atol=0.1)
 
     @repeat_test_std
     def test_fit_ellipsoid_straight(self):
         self.names = ('radius_z', 'radius_y', 'radius_x',
-                                'center_z', 'center_y', 'center_x')
+                      'center_z', 'center_y', 'center_x')
         self.atol = 1E-7
         radius = np.random.random((3)) * 10 + 5
         center = np.random.random((3)) * 10 + radius
@@ -116,13 +116,13 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
         radius_fit, center_fit, _ = fit_ellipsoid(points, mode='0')
 
         return np.concatenate([radius_fit, center_fit]), \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
 
     @repeat_test_std
     def test_fit_ellipsoid_straight_noisy(self):
         self.names = ('radius_z', 'radius_y', 'radius_x',
-                                'center_z', 'center_y', 'center_x')
+                      'center_z', 'center_y', 'center_x')
         self.atol = 0.1
         noise = 0.2
         radius = np.random.random((3)) * 10 + 5
@@ -133,7 +133,7 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
         radius_fit, center_fit, _ = fit_ellipsoid(points, mode='0')
 
         return np.concatenate([radius_fit, center_fit]), \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
     # def test_fit_ellipsoid_rotated(self):
     #     noise = 0.2
@@ -219,7 +219,7 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
         result = fit_ellipsoid(points, mode='xy', return_mode='skew')
 
         return np.concatenate(result), \
-               np.concatenate([radius, center, skew])
+            np.concatenate([radius, center, skew])
 
     @repeat_test_std
     def test_fit_ellipsoid_skewed_noisy(self):
@@ -253,7 +253,7 @@ class TestFits(RepeatedUnitTests, unittest.TestSuite):
         result = fit_ellipsoid(points, mode='xy', return_mode='skew')
 
         return np.concatenate(result), \
-               np.concatenate([radius, center, skew])
+            np.concatenate([radius, center, skew])
 
 
 class TestEllipse(RepeatedUnitTests, unittest.TestSuite):
@@ -268,7 +268,7 @@ class TestEllipse(RepeatedUnitTests, unittest.TestSuite):
         radius = (np.random.random() * 50 + 50,) * 2
         padding = [r + self.FWHM * 3 for r in radius]
         center = tuple([np.random.random() * (s - 2 * p) + p
-                       for (s, p) in zip(self.shape, padding)])
+                        for (s, p) in zip(self.shape, padding)])
         im = draw_ellipse(self.shape, radius, center,
                           FWHM=self.FWHM, noise=noise)
         result = find_ellipse(im)
@@ -283,13 +283,13 @@ class TestEllipse(RepeatedUnitTests, unittest.TestSuite):
         radius = (np.random.random() * 50 + 50,) * 2
         padding = [r + self.FWHM * 3 for r in radius]
         center = tuple([np.random.random() * (s - 2 * p) + p
-                       for (s, p) in zip(self.shape, padding)])
+                        for (s, p) in zip(self.shape, padding)])
         im = draw_ellipse(self.shape, radius, center,
                           FWHM=self.FWHM, noise=noise)
         result, _ = locate_ellipse(im)
 
         return result[['yr', 'xr', 'yc', 'xc']].values, \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
     @repeat_test_std
     def test_ellipse_noisy(self):
@@ -299,12 +299,12 @@ class TestEllipse(RepeatedUnitTests, unittest.TestSuite):
         radius = tuple(np.random.random(2) * 50 + 50,)
         padding = [r + self.FWHM * 3 for r in radius]
         center = tuple([np.random.random() * (s - 2 * p) + p
-                       for (s, p) in zip(self.shape, padding)])
+                        for (s, p) in zip(self.shape, padding)])
         im = draw_ellipse(self.shape, radius, center,
                           FWHM=self.FWHM, noise=noise)
         result, _ = locate_ellipse(im)
         return result[['yr', 'xr', 'yc', 'xc']].values, \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
 
 class TestEllipsoid(RepeatedUnitTests, unittest.TestSuite):
@@ -320,7 +320,7 @@ class TestEllipsoid(RepeatedUnitTests, unittest.TestSuite):
         radius = (R/ar, R, R)
         padding = [r + self.FWHM * 3 for r in radius]
         center = tuple([np.random.random() * (s - 2 * p) + p
-                       for (s, p) in zip(self.shape, padding)])
+                        for (s, p) in zip(self.shape, padding)])
         return center, radius
 
     @repeat_test_std
@@ -340,7 +340,7 @@ class TestEllipsoid(RepeatedUnitTests, unittest.TestSuite):
                             FWHM=self.FWHM, noise=0.02)
         result, _ = locate_ellipsoid_fast(im)
         return result[['zr', 'yr', 'xr', 'zc', 'yc', 'xc']].values, \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
     @repeat_test_std
     def test_locate_noisy(self):
@@ -350,129 +350,8 @@ class TestEllipsoid(RepeatedUnitTests, unittest.TestSuite):
                             FWHM=self.FWHM, noise=0.2)
         result, _ = locate_ellipsoid(im)
         return result[['zr', 'yr', 'xr', 'zc', 'yc', 'xc']].values, \
-               np.concatenate([radius, center])
+            np.concatenate([radius, center])
 
-
-class TestDisks(RepeatedUnitTests, unittest.TestCase):
-    """ Test case for finding circular disks """
-    repeats = 20
-    names = ('radius', 'y_coord', 'x_coord')
-
-    def generate_image(self, radius, n, noise=0.02):
-        """ Generate the test image """
-        image = SimulatedImage(shape=(300, 300), radius=radius,
-                               noise=noise)
-        image.draw_features(n, margin=2*radius, separation=2*radius + 2)
-        return image
-
-    @repeat_test_std
-    def test_find_single(self):
-        """ Test finding single particle """
-        self.atol = 5
-        radius = np.random.uniform(15, 30)
-        generated_image = self.generate_image(radius, 1)
-
-        fits = find_disks(generated_image.image, (radius / 2.0,
-                                                  radius * 2.0),
-                          number_of_disks=1)
-
-        y_coord, x_coord = generated_image.coords[0]
-        if len(fits) != 1:  # Particle number mismatch
-            r, y, x = np.nan, np.nan, np.nan
-        else:
-            r, x, y = fits[['r', 'x', 'y']].values[0]
-
-        return (r, y, x), (radius, y_coord, x_coord)
-
-    @repeat_test_std
-    def test_find_single_noisy(self):
-        """ Test find single noisy particle """
-        self.atol = 5
-        radius = np.random.uniform(15, 30)
-        generated_image = self.generate_image(radius, 1, noise=0.2)
-
-        fits = find_disks(generated_image.image, (radius / 2.0,
-                                                  radius * 2.0),
-                          number_of_disks=1)
-
-        y_coord, x_coord = generated_image.coords[0]
-        if len(fits) != 1: # Particle number mismatch
-            r, y, x = np.nan, np.nan, np.nan
-        else:
-            r, x, y = fits[['r', 'x', 'y']].values[0]
-
-        return (r, y, x), (radius, y_coord, x_coord)
-
-    @repeat_test_std
-    def test_find_multiple_noisy(self):
-        """ Test finding multiple particles (noisy) """
-        self.atol = 5
-        radius = np.random.uniform(15, 30)
-        generated_image = self.generate_image(radius, 10, noise=0.2)
-        actual_number = len(generated_image.coords)
-        fits = find_disks(generated_image.image, (radius / 2.0,
-                                                  radius * 2.0),
-                          number_of_disks=actual_number)
-
-        _, coords = sort_positions(generated_image.coords,
-                                   np.array([fits['y'].values,
-                                             fits['x'].values]).T)
-
-        if len(fits) == 0:  # Nothing found
-            actual = np.repeat([[np.nan, np.nan, np.nan]], actual_number,
-                               axis=0)
-        else:
-            actual = fits[['r', 'y', 'x']].values.astype(np.float64)
-
-        expected = np.array([np.full(actual_number, radius, np.float64),
-                             coords[:, 0], coords[:, 1]]).T
-
-        return np.sqrt(((actual - expected)**2).mean(0)), [0] * 3
-
-    @repeat_test_std
-    def test_locate_single_noisy(self):
-        """ Test locating single particle (noisy) """
-        self.atol = 0.5
-        radius = np.random.uniform(15, 30)
-        generated_image = self.generate_image(radius, 1, noise=0.2)
-
-        fits = locate_disks(generated_image.image, (radius / 2.0,
-                                                    radius * 2.0),
-                            number_of_disks=1)
-
-        y_coord, x_coord = generated_image.coords[0]
-        if len(fits) != 1:  # Particle number mismatch
-            r, y, x = np.nan, np.nan, np.nan
-        else:
-            r, y, x = fits[['r', 'y', 'x']].values[0]
-
-        return (r, y, x), (radius, y_coord, x_coord)
-
-    @repeat_test_std
-    def test_locate_multiple_noisy(self):
-        """ Test locating multiple particles (noisy) """
-        self.atol = 0.5
-        radius = np.random.uniform(15, 30)
-        generated_image = self.generate_image(radius, 10, noise=0.2)
-        actual_number = len(generated_image.coords)
-        fits = locate_disks(generated_image.image, (radius / 2.0,
-                                                    radius * 2.0),
-                            number_of_disks=actual_number)
-
-        _, coords = sort_positions(generated_image.coords,
-                                   np.array([fits['y'].values,
-                                             fits['x'].values]).T)
-
-        if len(fits) == 0:  # Nothing found
-            actual = np.repeat([[np.nan, np.nan, np.nan]], actual_number,
-                               axis=0)
-        else:
-            actual = fits[['r', 'y', 'x']].values.astype(np.float64)
-
-        expected = np.array([np.full(actual_number, radius, np.float64),
-                             coords[:, 0], coords[:, 1]]).T
-
-        return np.sqrt(((actual - expected)**2).mean(0)), [0] * 3
 
 if __name__ == '__main__':
     nose.runmodule(argv=[__file__, '-vvs', '-x'], exit=False)
